smart-city-traffic-analysis/airflow/requirements.txt
Flask-Session
apache-airflow==2.8.1
pandas
requests
google-cloud-bigquery
